+++ src/kodo_python/decoder.hpp
@@ -41,6 +41,22 @@ namespace kodo_python
     }
 
     template<class Decoder>
+    PyObject* copy_from_symbol(Decoder& decoder, uint32_t index)
+    {
+        std::vector<uint8_t> payload(decoder.symbol_size());
+        auto storage = sak::mutable_storage(
+            payload.data(), decoder.symbol_size());
+        decoder.copy_from_symbol(index, storage);
+        #if PY_MAJOR_VERSION >= 3
+        return PyBytes_FromStringAndSize(
+            (char*)payload.data(), decoder.symbol_size());
+        #else
+        return PyString_FromStringAndSize(
+            (char*)payload.data(), decoder.symbol_size());
+        #endif
+    }
+
+    template<class Decoder>
     PyObject* decoder_write_payload(Decoder& decoder)
     {
         std::vector<uint8_t> payload(decoder.payload_size());
@@ -151,13 +167,54 @@ namespace kodo_python
             "performance reasons, choose to not keep track of the exact\n"
             "status of the decoding matrix.\n"
             "It is however guaranteed that at least this amount of uncoded\n"
-            "symbols exist.\n"
+            "symbols exist.\n\n"
             "\t:returns: The number of symbols which have been uncoded.\n"
+        )
+        .def("symbols_missing", &decoder_type::symbols_missing,
+            "Return the number of missing symbols at the decoder.\n\n"
+            "\t:returns: The number of missing symbols.\n"
+        )
+        .def("symbols_partially_decoded",
+            &decoder_type::symbols_partially_decoded,
+            "Return the number of partially decoded symbols at the decoder.\n\n"
+            "\t:returns: The number of partially decoded symbols.\n"
+        )
+        .def("is_symbol_uncoded", &decoder_type::is_symbol_uncoded,
+            arg("index"),
+            "Check if the symbol at given index is uncoded.\n\n"
+            "This may return false for symbols that are actually uncoded,\n"
+            "but never true for symbols that are not uncoded.\n"
+            "As with the symbols_uncoded() function the reason for this is\n"
+            "that some algorithms do not, for performance reasons, keep track\n"
+            "of the exact status of the decoding matrix.\n\n"
+            "\t:param index: Index of the symbol to check.\n"
+            "\t:return: True if the symbol is uncoded, and otherwise false.\n"
+        )
+        .def("is_symbol_missing", &decoder_type::is_symbol_missing,
+            arg("index"),
+            "Check if the symbol at given index is missing.\n\n"
+            "\t:param index: Index of the symbol to check.\n"
+            "\t:return: True if the symbol is missing otherwise false.\n"
+        )
+        .def("is_symbol_partially_decoded",
+            &decoder_type::is_symbol_partially_decoded,
+            arg("index"),
+            "Check if the symbol at given index is partially decoded.\n\n"
+            "\t:param index: Index of the symbol to check.\n"
+            "\t:return: True if the symbol is partially decoded otherwise\n"
+            "\t         false.\n"
+        )
+        .def("copy_from_symbol", &copy_from_symbol<decoder_type>,
+            arg("index"),
+            "Return the decoded symbol.\n\n"
+            "\t:param index: Index of the symbol to return.\n"
+            "\t:returns: The decoded symbol.\n"
         )
         .def("copy_from_symbols", &copy_from_symbols<decoder_type>,
             "Return the decoded symbols.\n\n"
             "\t:returns: The decoded symbols.\n"
         );
+
 
         (write_payload_method<
             kodo_core::has_write_payload<decoder_type>::value>(decoder_class));