+++ NEWS.rst
@@ -6,7 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added decoder functions: ``copy_from_symbol``, ``is_symbol_missing``,
+  ``is_symbol_partially_decoded``, ``is_symbol_uncoded``, ``symbols_missing``,
+  ``symbols_partially_decoded``.
+* Patch: Updated documentation.
 
 10.0.0
 ------